generic constraint commitments
